+++ setup.cfg
@@ -8,7 +8,7 @@ ignore = C901,E0602,E501,W605
 [tool:pytest]
 
 [coverage:run]
-omit = main.py, */__init__.py, visma/gui/*, visma/testbed/*
+omit = main.py, setup.py, */__init__.py, visma/gui/*, visma/testbed/*
 
 [coverage:report]
-omit = main.py, */__init__.py, visma/gui/*, visma/testbed/*
+omit = main.py, setup.py, */__init__.py, visma/gui/*, visma/testbed/*
