+++ setup.py
@@ -4,9 +4,9 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="visma",
-    description="VISual MAth - A math equation solver and visualizer",
-    version="0.2.0",
+    name="VISualMAth",
+    description="visma - VISual MAth : A math equation solver and visualizer",
+    version="0.2.1",
     author="Siddharth Kothiyal, Shantanu Mishra",
     author_email="sid.kothiyal27@gmail.com, 8hantanu@gmail.com",
     long_description=long_description,
@@ -18,7 +18,6 @@ setuptools.setup(
         'Issues': 'https://github.com/aerospaceresearch/visma/issues',
         'Chat': 'https://gitter.im/aerospaceresearch/visma'
     },
-    license="GPLv3",
     packages=setuptools.find_packages(),
     classifiers=(
         "Programming Language :: Python :: 3",
@@ -27,6 +26,5 @@ setuptools.setup(
         "Topic :: Scientific/Engineering",
         "Topic :: Scientific/Engineering :: Mathematics",
         "Topic :: Scientific/Engineering :: Visualization"
-    ),
-    python_requires='>=3'
+    )
 )
