+++ visma/config/define.py
